same as the last commit

+++ ax_fft/bellerophon/bellerophon_functions.cpp
@@ -49,7 +49,7 @@ extern "C" double BELLERO_getError(){
 		count++;
 	}
 	
-	avgError = absError/count;
+	avgError = (double)absError/count;
 
 	return avgError;
 }
